compatibility: rename namespace from `Daht.Sagitta.Core.Monads` to `Daht.Sagitta.Core.Results`
