+++ libraries/core/source/Global.cs
@@ -8,5 +8,5 @@ global using System.Diagnostics.CodeAnalysis;
 global using System.Diagnostics.Contracts;
 global using System.Runtime.CompilerServices;
 global using System.Runtime.InteropServices;
-global using Daht.Sagitta.Core.Monads.Exceptions.Helpers;
+global using Daht.Sagitta.Core.Results.Exceptions.Helpers;
 global using Daht.Sagitta.Shared.Analysis;
