+++ libraries/core/tests/unit/Global.cs
@@ -6,11 +6,11 @@
 global using System.Diagnostics.CodeAnalysis;
 global using System.Globalization;
 global using System.Reflection;
-global using Daht.Sagitta.Core.Monads;
-global using Daht.Sagitta.Core.UnitTests.Monads.Asserters;
-global using Daht.Sagitta.Core.UnitTests.Monads.Fakers;
-global using Daht.Sagitta.Core.UnitTests.Monads.Fixtures;
-global using Daht.Sagitta.Core.UnitTests.Monads.Mothers;
+global using Daht.Sagitta.Core.Results;
+global using Daht.Sagitta.Core.UnitTests.Results.Asserters;
+global using Daht.Sagitta.Core.UnitTests.Results.Fakers;
+global using Daht.Sagitta.Core.UnitTests.Results.Fixtures;
+global using Daht.Sagitta.Core.UnitTests.Results.Mothers;
 global using Daht.Sagitta.Shared.Analysis;
 global using Daht.Sagitta.Shared.UnitTests.Exceptions;
 global using Xunit;
